+++ include/metashell/header_file_environment.hpp
@@ -38,8 +38,6 @@ namespace metashell
     virtual std::string get() const override;
     virtual std::string get_appended(const std::string& s_) const override;
 
-    virtual boost::filesystem::path internal_dir() const override;
-
     virtual const data::headers& get_headers() const override;
 
     virtual std::string get_all() const override;