+++ lib/core/engine_entry.cpp
@@ -30,13 +30,14 @@ engine_entry::engine_entry(engine_factory factory_,
 std::unique_ptr<iface::engine>
 engine_entry::build(const data::config& config_,
                     const boost::filesystem::path& internal_dir_,
+                    const boost::filesystem::path& temp_dir_,
                     const boost::filesystem::path& env_filename_,
                     iface::environment_detector& env_detector_,
                     iface::displayer& displayer_,
                     logger* logger_) const
 {
-  return _factory(config_, internal_dir_, env_filename_, env_detector_,
-                  displayer_, logger_);
+  return _factory(config_, internal_dir_, temp_dir_, env_filename_,
+                  env_detector_, displayer_, logger_);
 }
 
 const std::string& engine_entry::args() const { return _args; }