+++ test/unit/breaking_environment.cpp
@@ -79,11 +79,6 @@ void breaking_environment::get_appended_throw_from_now()
   _get_appended_throw = true;
 }
 
-boost::filesystem::path breaking_environment::internal_dir() const
-{
-  return get_headers().internal_dir();
-}
-
 const data::headers& breaking_environment::get_headers() const
 {
   return _headers;