+++ test/unit/test_breaking_environment.cpp
@@ -29,7 +29,6 @@ JUST_TEST_CASE(test_breaking_environment_does_not_throw)
   e.append("foo");
   e.get();
   e.get_appended("bar");
-  e.internal_dir();
   e.get_headers();
 }
 
@@ -44,7 +43,6 @@ JUST_TEST_CASE(test_breaking_environment_append_throws)
                      });
   e.get();
   e.get_appended("bar");
-  e.internal_dir();
   e.get_headers();
 }
 
@@ -59,6 +57,5 @@ JUST_TEST_CASE(test_breaking_environment_get_appended_throws)
                      {
                        e.get_appended("bar");
                      });
-  e.internal_dir();
   e.get_headers();
 }